[feat] social Login Manager 추가사항 리펙토링 (#109)
